CMS-1132: Update re sync all script
