+++ backend/strapi-sync/reset-and-import-data.js
@@ -36,12 +36,13 @@ export async function deleteAllData() {
   await User.destroy({ where: {} });
 }
 
+// TODO: clean up this function
 export async function importData() {
   await syncData();
-  await oneTimeDataImport();
+  // await oneTimeDataImport();
   await createSingleItemsCampgrounds();
   await createMultipleItemsCampgrounds();
-  await createMissingDatesAndSeasons();
+  // await createMissingDatesAndSeasons();
 }
 
 export async function resetScript() {
