+++ backend/tasks/re-sync-all/delete-all.js
